+++ lib/http/authorization_header/bearer_token.rb
@@ -4,18 +4,17 @@ module HTTP
   module AuthorizationHeader
     # OAuth2 Bearer token authorization header builder
     # @see http://tools.ietf.org/html/rfc6750
+    #
+    # @deprecated Will be remove in v0.8.0
     class BearerToken
       # @param [#fetch] opts
       # @option opts [#to_s] :token
       # @option opts [Boolean] :encode (false) deprecated
       def initialize(opts)
+        warn "#{Kernel.caller.first}: [DEPRECATION] BearerToken deprecated."
+
         @token = opts.fetch :token
-
-        return unless opts.fetch(:encode, false)
-
-        warn "#{Kernel.caller.first}: [DEPRECATION] BearerToken :encode option is deprecated. " \
-          "You should pass encoded token on your own: { :token => Base64.strict_encode64('token') }"
-        @token = Base64.strict_encode64 @token
+        @token = Base64.strict_encode64 @token if opts.fetch(:encode, false)
       end
 
       # :nodoc: