+++ CHANGES.md
@@ -1,7 +1,7 @@
 0-6-stable (unreleased)
 -----------------------
 
-* Deprecate `:encode` option BearerToken authorization header. (@ixti)
+* Deprecate BearerToken authorization header. (@ixti)
 * Fix handling of chunked responses without Content-Length header. (@ixti)
 * Rename `HTTP.with_follow` to `HTTP.follow` and mark former one as being
   deprecated (@ixti)